first menu while and 3. exit

+++ EcoEnergySolutions/HelperClass.cs
@@ -8,7 +8,7 @@ namespace EcoEnergySolutions
 {
     public class HelperClass
     {
-        public static void BucleComprovacio(string Missatge, float num)
+        public static void BucleComprovacio(string Missatge, int num)
         {
             bool flag = true;
             while (flag)
